python_pricechartingscraper/requirements.txt
beautifulsoup4>=4.9.3
requests>=2.26.0
PyYAML>=5.4.1
python-dateutil>=2.8.2
Pillow>=10.0.0  # For image processing and WebP conversion 